narrow api
